Update localstack version in docker-compose
This change updates the localstack version within the docker-compose
files to the latest 1.4.0 version.

There are significant changes included in this version upgrade:
- Changes to healthcheck endpoint as previous is deprecated
- Removal of `DEFAULT_REGION` as deprecated
- Change to init folder as previous is deprecated.

Other changes include:
- Marking scripts with executable permissions
- Change to service check to regex against either `available` and
  `running`.

This change also includes moving to OpenSearch from ElasticSearch. The
domain is created with version 1.3, which we are running in
environments. To support this, a better domain ready check has been
added.